Fixed atomic operation bug on ARMv7-M.
Fixed an inline assembly bug on the ARMv7-M. For the strex instruction,
Rd can not be the same as Rt or Rn (Read the manual, Ye!). We had Rd and
Rn both as R0 in the inline assembly. But somehow, STM32F4 was fairly
tolerant and did not panic on it. However, the new STM32F7 generated a
UsageFault with undefined instruction status.