Bug fix for AVR.
Fixed a bug that set the initial task SREG to enable interrupt. This is
wrong because switching to a new task is treated as full context switch.
And this assumes task is in kernel and interrupt is disabled. Also
removed the stack switch in McuInit. We will just use the stack when
StartOS is called temporarily until switching to the first user task.